Fix: unificacion del form_ticket.py con el model ticket
